factrorisation in scenario_test
