+++ integration/scenario_test.go
@@ -11,7 +11,6 @@ import (
 	"math/rand"
 	"net/http"
 	"os"
-	"os/exec"
 	"reflect"
 	"strconv"
 	"strings"
@@ -45,7 +44,7 @@ func TestScenario(t *testing.T) {
 		numNodes, err = strconv.Atoi(n)
 		require.NoError(t, err)
 	}
-	t.Run("Basic configuration", getScenarioTest(numNodes, numNodes, 1))
+	t.Run("Basic configuration", getScenarioTest(numNodes, 20, 1))
 }
 
 func getScenarioTest(numNodes int, numVotes int, numForm int) func(*testing.T) {
@@ -86,92 +85,47 @@ func startFormProcess(wg *sync.WaitGroup, numNodes int, numVotes int, proxyArray
 	secret := suite.Scalar()
 	err = secret.UnmarshalBinary(secretkeyBuf)
 	require.NoError(t, err)
+	timeTable := make([]float64, 10)
+
+	step := 0
 
 	// ###################################### CREATE SIMPLE FORM ######
+	oldTime := time.Now()
 
-	t.Log("Create form")
+	formID := createFormScenario(contentType, proxyArray[0], secret, t)
 
-	configuration := fake.BasicConfiguration
+	timeTable[step] = time.Since(oldTime).Seconds()
+	t.Logf("Creating the election takes: %v sec", timeTable[step])
 
-	createSimpleFormRequest := ptypes.CreateFormRequest{
-		Configuration: configuration,
-		AdminID:       "adminId",
-	}
-
-	signed, err := createSignedRequest(secret, createSimpleFormRequest)
-	require.NoError(t, err)
-
-	resp, err := http.Post(proxyArray[0]+"/evoting/forms", contentType, bytes.NewBuffer(signed))
-	require.NoError(t, err)
-	body, err := io.ReadAll(resp.Body)
-	require.NoError(t, err)
-
-	require.Equal(t, resp.StatusCode, http.StatusOK, "unexpected status: %s", body)
-
-	t.Log("response body:", string(body))
-	resp.Body.Close()
-
-	var createFormResponse ptypes.CreateFormResponse
-
-	err = json.Unmarshal(body, &createFormResponse)
-	require.NoError(t, err)
-
-	formID := createFormResponse.FormID
-
-	ok, err := pollTxnInclusion(proxyArray[0], createFormResponse.Token, t)
-	require.NoError(t, err)
-	require.True(t, ok)
-
-	t.Logf("ID of the form : " + formID)
+	step++
 
 	// ##################################### SETUP DKG #########################
+	oldTime = time.Now()
 
-	t.Log("Init DKG")
+	startDKGScenario(numNodes, timeTable, formID, secret, proxyArray, t)
 
-	for i := 0; i < numNodes; i++ {
-		t.Log("Node" + strconv.Itoa(i+1))
-		t.Log(proxyArray[i])
-		err = initDKG(secret, proxyArray[i], formID, t)
-		require.NoError(t, err)
+	timeTable[step] = time.Since(oldTime).Seconds()
+	t.Logf("DKG setup takes: %v sec", timeTable[step])
 
-	}
-	t.Log("Setup DKG")
-
-	msg := ptypes.UpdateDKG{
-		Action: "setup",
-	}
-	signed, err = createSignedRequest(secret, msg)
-	require.NoError(t, err)
-
-	req, err := http.NewRequest(http.MethodPut, proxyArray[0]+"/evoting/services/dkg/actors/"+formID, bytes.NewBuffer(signed))
-	require.NoError(t, err)
-
-	t.Log("Sending")
-	resp, err = http.DefaultClient.Do(req)
-	require.NoError(t, err)
-	require.Equal(t, resp.StatusCode, http.StatusOK, "unexpected status: %s", resp.Status)
+	step++
 
 	// ##################################### OPEN FORM #####################
 	t.Log("Open form")
-	// Wait for DKG setup
-	timeTable := make([]float64, 5)
-	oldTime := time.Now()
-
-	err = waitForDKG(proxyArray[0], formID, time.Second*100, t)
-	require.NoError(t, err)
-
-	currentTime := time.Now()
-	diff := currentTime.Sub(oldTime)
-	timeTable[0] = diff.Seconds()
-	t.Logf("DKG setup takes: %v sec", diff.Seconds())
 
 	randomproxy := "http://localhost:9081"
 	t.Logf("Open form send to proxy %v", randomproxy)
 
-	t.Log("Open form")
-	ok, err = updateForm(secret, randomproxy, formID, "open", t)
+	oldTime = time.Now()
+
+	ok, err := updateForm(secret, randomproxy, formID, "open", t)
 	require.NoError(t, err)
 	require.True(t, ok)
+
+	timeTable[step] = time.Since(oldTime).Seconds()
+	t.Logf("Opening the election takes: %v sec", timeTable[step])
+
+	step++
+
 	// ##################################### GET FORM INFO #################
 
 	proxyAddr1 := proxyArray[0]
@@ -197,6 +151,8 @@ func startFormProcess(wg *sync.WaitGroup, numNodes int, numVotes int, proxyArray
 	// ##################################### CAST BALLOTS ######################
 
 	t.Log("cast ballots")
+
+	oldTime = time.Now()
 
 	//make List of ballots
 	b1 := string("select:" + encodeIDBallot("bb") + ":0,0,1,0\n" + "text:" + encodeIDBallot("ee") + ":eWVz\n\n") //encoding of "yes"
@@ -226,12 +182,9 @@ func startFormProcess(wg *sync.WaitGroup, numNodes int, numVotes int, proxyArray
 	}
 
 	for i := 0; i < numVotes; i++ {
-		// t.Logf("ballot in str is: %v", ballotList[i])
 
 		ballot, err := marshallBallotManual(ballotList[i], pubKey, Chunksperballot)
 		require.NoError(t, err)
-
-		// t.Logf("ballot is: %v", ballot)
 
 		castVoteRequest := ptypes.CastVoteRequest{
 			UserID: "user" + strconv.Itoa(i+1),
@@ -241,15 +194,14 @@ func startFormProcess(wg *sync.WaitGroup, numNodes int, numVotes int, proxyArray
 		randomproxy = proxyArray[rand.Intn(len(proxyArray))]
 		t.Logf("cast ballot to proxy %v", randomproxy)
 
-		// t.Logf("vote is: %v", castVoteRequest)
-		signed, err = createSignedRequest(secret, castVoteRequest)
+		signed, err := createSignedRequest(secret, castVoteRequest)
 		require.NoError(t, err)
 
-		resp, err = http.Post(randomproxy+"/evoting/forms/"+formID+"/vote", contentType, bytes.NewBuffer(signed))
+		resp, err := http.Post(randomproxy+"/evoting/forms/"+formID+"/vote", contentType, bytes.NewBuffer(signed))
 		require.NoError(t, err)
 		require.Equal(t, http.StatusOK, resp.StatusCode, "unexpected status: %s", resp.Status)
 
-		body, err = io.ReadAll(resp.Body)
+		body, err := io.ReadAll(resp.Body)
 		require.NoError(t, err)
 
 		var infos ptypes.TransactionInfoToSend
@@ -263,6 +215,12 @@ func startFormProcess(wg *sync.WaitGroup, numNodes int, numVotes int, proxyArray
 		resp.Body.Close()
 
 	}
+
+	timeTable[step] = time.Since(oldTime).Seconds()
+	t.Logf("Casting %v ballots takes: %v sec", numVotes, timeTable[step])
+
+	step++
+
 	time.Sleep(time.Second * 5)
 
 	// Kill and restart the node, change false to true when we want to use
@@ -282,6 +240,9 @@ func startFormProcess(wg *sync.WaitGroup, numNodes int, numVotes int, proxyArray
 	}
 
 	// ############################# CLOSE FORM FOR REAL ###################
+
+	oldTime = time.Now()
+
 	randomproxy = proxyArray[rand.Intn(len(proxyArray))]
 
 	t.Logf("Close form (for real) send to proxy %v", randomproxy)
@@ -289,6 +250,11 @@ func startFormProcess(wg *sync.WaitGroup, numNodes int, numVotes int, proxyArray
 	ok, err = updateForm(secret, randomproxy, formID, "close", t)
 	require.NoError(t, err)
 	require.True(t, ok)
+
+	timeTable[step] = time.Since(oldTime).Seconds()
+	t.Logf("Closing form takes: %v sec", timeTable[step])
+
+	step++
 
 	time.Sleep(time.Second * 3)
 
@@ -303,32 +269,34 @@ func startFormProcess(wg *sync.WaitGroup, numNodes int, numVotes int, proxyArray
 
 	t.Log("shuffle ballots")
 
+	oldTime = time.Now()
+
 	shuffleBallotsRequest := ptypes.UpdateShuffle{
 		Action: "shuffle",
 	}
 
-	signed, err = createSignedRequest(secret, shuffleBallotsRequest)
+	signed, err := createSignedRequest(secret, shuffleBallotsRequest)
 	require.NoError(t, err)
 
 	randomproxy = proxyArray[rand.Intn(len(proxyArray))]
 
-	req, err = http.NewRequest(http.MethodPut, randomproxy+"/evoting/services/shuffle/"+formID, bytes.NewBuffer(signed))
+	req, err := http.NewRequest(http.MethodPut, randomproxy+"/evoting/services/shuffle/"+formID, bytes.NewBuffer(signed))
+	require.NoError(t, err)
+
+	resp, err := http.DefaultClient.Do(req)
 	require.NoError(t, err)
 	require.Equal(t, resp.StatusCode, http.StatusOK, "unexpected status: %s", resp.Status)
 
-	resp, err = http.DefaultClient.Do(req)
-	require.NoError(t, err)
-
-	currentTime = time.Now()
-	diff = currentTime.Sub(oldTime)
-	timeTable[1] = diff.Seconds()
-	t.Logf("Shuffle takes: %v sec", diff.Seconds())
-
-	body, err = io.ReadAll(resp.Body)
+	body, err := io.ReadAll(resp.Body)
 	require.NoError(t, err)
 
 	t.Log("Response body: " + string(body))
 	resp.Body.Close()
+
+	timeTable[step] = time.Since(oldTime).Seconds()
+	t.Logf("Shuffling ballots takes: %v sec", timeTable[step])
+
+	step++
 
 	getFormResponse = getFormInfo(proxyAddr1, formID, t)
 	formStatus = getFormResponse.Status
@@ -345,30 +313,24 @@ func startFormProcess(wg *sync.WaitGroup, numNodes int, numVotes int, proxyArray
 	t.Log("request public shares")
 
 	randomproxy = proxyArray[rand.Intn(len(proxyArray))]
+
 	oldTime = time.Now()
 
 	_, err = updateDKG(secret, randomproxy, formID, "computePubshares", t)
 	require.NoError(t, err)
 
-	currentTime = time.Now()
-	diff = currentTime.Sub(oldTime)
-	timeTable[2] = diff.Seconds()
+	timeTable[step] = time.Since(oldTime).Seconds()
+	t.Logf("Requesting public shares takes: %v sec", timeTable[step])
 
-	t.Logf("Request public share takes: %v sec", diff.Seconds())
+	step++
 
-	time.Sleep(10 * time.Second)
+	time.Sleep(7 * time.Second)
 
 	getFormResponse = getFormInfo(proxyAddr1, formID, t)
 	formStatus = getFormResponse.Status
 
-	oldTime = time.Now()
 	err = waitForFormStatus(proxyAddr1, formID, uint16(4), time.Second*300, t)
 	require.NoError(t, err)
-
-	currentTime = time.Now()
-	diff = currentTime.Sub(oldTime)
-	timeTable[4] = diff.Seconds()
-	t.Logf("Status goes to 4 takes: %v sec", diff.Seconds())
 
 	t.Logf("Status of the form : %v", formStatus)
 	require.Equal(t, uint16(4), formStatus)
@@ -379,17 +341,17 @@ func startFormProcess(wg *sync.WaitGroup, numNodes int, numVotes int, proxyArray
 	t.Log("decrypt ballots")
 
 	randomproxy = proxyArray[rand.Intn(len(proxyArray))]
+
 	oldTime = time.Now()
 
 	ok, err = updateForm(secret, randomproxy, formID, "combineShares", t)
 	require.NoError(t, err)
 	require.True(t, ok)
 
-	currentTime = time.Now()
-	diff = currentTime.Sub(oldTime)
-	timeTable[3] = diff.Seconds()
+	timeTable[step] = time.Since(oldTime).Seconds()
+	t.Logf("Decrypting ballots takes: %v sec", timeTable[step])
 
-	t.Logf("decryption takes: %v sec", diff.Seconds())
+	step++
 
 	time.Sleep(time.Second * 3)
 
@@ -424,11 +386,15 @@ func startFormProcess(wg *sync.WaitGroup, numNodes int, numVotes int, proxyArray
 	}
 
 	require.True(t, tmpCount, "front end votes are different from decrypted votes")
-	t.Logf("DKG setup time : %v", timeTable[0])
-	t.Logf("shuffle time : %v", timeTable[1])
-	t.Logf("Public share time : %v", timeTable[2])
-	t.Logf("Status goes to 4 takes: %v sec", diff.Seconds())
-	t.Logf("decryption time : %v", timeTable[3])
+	t.Logf("Creating the form takes %v sec", timeTable[0])
+	t.Logf("Setting up DKG takes %v sec", timeTable[1])
+	t.Logf("Oppening the form takes %v sec", timeTable[2])
+	t.Logf("Casting %v ballots takes %v sec", numVotes, timeTable[3])
+	t.Logf("Closing the form takes %v sec", timeTable[4])
+	t.Logf("Shuffling ballots takes %v sec", timeTable[5])
+	t.Logf("Requesting public shares takes %v sec", timeTable[6])
+	t.Logf("Decrypting ballots takes %v sec", timeTable[7])
+
 }
 
 // -----------------------------------------------------------------------------
@@ -528,182 +494,4 @@ func createSignedRequest(secret kyber.Scalar, msg interface{}) ([]byte, error) {
 	}
 
 	return signedJSON, nil
-}
-
-func initDKG(secret kyber.Scalar, proxyAddr, formIDHex string, t *testing.T) error {
-	setupDKG := ptypes.NewDKGRequest{
-		FormID: formIDHex,
-	}
-
-	signed, err := createSignedRequest(secret, setupDKG)
-	require.NoError(t, err)
-
-	resp, err := http.Post(proxyAddr+"/evoting/services/dkg/actors", "application/json", bytes.NewBuffer(signed))
-	if err != nil {
-		return xerrors.Errorf("failed to post request: %v", err)
-	}
-	require.Equal(t, resp.StatusCode, http.StatusOK, "unexpected status: %s", resp.Status)
-
-	return nil
-}
-
-func updateForm(secret kyber.Scalar, proxyAddr, formIDHex, action string, t *testing.T) (bool, error) {
-	msg := ptypes.UpdateFormRequest{
-		Action: action,
-	}
-
-	signed, err := createSignedRequest(secret, msg)
-	require.NoError(t, err)
-
-	req, err := http.NewRequest(http.MethodPut, proxyAddr+"/evoting/forms/"+formIDHex, bytes.NewBuffer(signed))
-	if err != nil {
-		return false, xerrors.Errorf("failed to create request: %v", err)
-	}
-
-	resp, err := http.DefaultClient.Do(req)
-	if err != nil {
-		return false, xerrors.Errorf("failed retrieve the decryption from the server: %v", err)
-	}
-	body, err := io.ReadAll(resp.Body)
-	if err != nil {
-		return false, xerrors.Errorf("failed to read response body: %v", err)
-	}
-	require.Equal(t, resp.StatusCode, http.StatusOK, "unexpected status: %s", body)
-
-	//use the pollTxnInclusion func
-	var result map[string]interface{}
-	err = json.Unmarshal(body, &result)
-	if err != nil {
-		return false, xerrors.Errorf("failed to unmarshal response body: %v", err)
-	}
-
-	return pollTxnInclusion(proxyAddr, result["Token"].(string), t)
-
-}
-
-func updateDKG(secret kyber.Scalar, proxyAddr, formIDHex, action string, t *testing.T) (int, error) {
-	msg := ptypes.UpdateDKG{
-		Action: action,
-	}
-
-	signed, err := createSignedRequest(secret, msg)
-	require.NoError(t, err)
-
-	req, err := http.NewRequest(http.MethodPut, proxyAddr+"/evoting/services/dkg/actors/"+formIDHex, bytes.NewBuffer(signed))
-	if err != nil {
-		return 0, xerrors.Errorf("failed to create request: %v", err)
-	}
-
-	resp, err := http.DefaultClient.Do(req)
-	if err != nil {
-		return 0, xerrors.Errorf("failed to execute the query: %v", err)
-	}
-
-	require.Equal(t, resp.StatusCode, http.StatusOK, "unexpected status: %s", resp.Status)
-
-	return 0, nil
-}
-
-func getFormInfo(proxyAddr, formID string, t *testing.T) ptypes.GetFormResponse {
-	// t.Log("Get form info")
-
-	resp, err := http.Get(proxyAddr + "/evoting/forms" + "/" + formID)
-	require.NoError(t, err)
-
-	var infoForm ptypes.GetFormResponse
-	decoder := json.NewDecoder(resp.Body)
-
-	err = decoder.Decode(&infoForm)
-	require.NoError(t, err)
-
-	resp.Body.Close()
-
-	return infoForm
-
-}
-
-func killNode(proxyArray []string, nodeNub int, t *testing.T) []string {
-
-	proxyArray[nodeNub-1] = proxyArray[len(proxyArray)-1]
-	proxyArray[len(proxyArray)-1] = ""
-	proxyArray = proxyArray[:len(proxyArray)-1]
-
-	cmd := exec.Command("docker", "kill", fmt.Sprintf("node%v", nodeNub))
-	err := cmd.Run()
-	require.NoError(t, err)
-
-	return proxyArray
-}
-
-func restartNode(nodeNub int, t *testing.T) {
-	cmd := exec.Command("docker", "restart", fmt.Sprintf("node%v", nodeNub))
-	err := cmd.Run()
-	require.NoError(t, err)
-
-	// Replace the relative path
-	cmd = exec.Command("rm", fmt.Sprintf("/Users/jean-baptistezhang/EPFL_cours/semestre_2/d-voting/nodedata/node%v/daemon.sock", nodeNub))
-	err = cmd.Run()
-	require.NoError(t, err)
-
-	cmd = exec.Command("bash", "-c", fmt.Sprintf("docker exec -d node%v memcoin --config /tmp/node%v start --postinstall --promaddr :9100 --proxyaddr :9080 --proxykey adbacd10fdb9822c71025d6d00092b8a4abb5ebcb673d28d863f7c7c5adaddf3 --listen tcp://0.0.0.0:2001 --public //172.18.0.%v:2001", nodeNub, nodeNub, nodeNub+1))
-	err = cmd.Run()
-	require.NoError(t, err)
-}
-
-func getDKGInfo(proxyAddr, formID string, t *testing.T) ptypes.GetActorInfo {
-
-	resp, err := http.Get(proxyAddr + "/evoting/services/dkg/actors" + "/" + formID)
-	require.NoError(t, err)
-
-	var infoDKG ptypes.GetActorInfo
-	decoder := json.NewDecoder(resp.Body)
-
-	err = decoder.Decode(&infoDKG)
-	require.NoError(t, err)
-
-	resp.Body.Close()
-
-	return infoDKG
-
-}
-
-func waitForDKG(proxyAddr, formID string, timeOut time.Duration, t *testing.T) error {
-	expired := time.Now().Add(timeOut)
-
-	isOK := func() bool {
-		infoDKG := getDKGInfo(proxyAddr, formID, t)
-		t.Logf("DKG info: %+v", infoDKG)
-
-		return infoDKG.Status == 1
-	}
-
-	for !isOK() {
-
-		if time.Now().After(expired) {
-			return xerrors.New("expired")
-		}
-
-		time.Sleep(time.Millisecond * 1000)
-	}
-
-	return nil
-}
-
-func waitForFormStatus(proxyAddr, formID string, status uint16, timeOut time.Duration, t *testing.T) error {
-	expired := time.Now().Add(timeOut)
-
-	isOK := func() bool {
-		infoForm := getFormInfo(proxyAddr, formID, t)
-		return infoForm.Status == status
-	}
-
-	for !isOK() {
-		if time.Now().After(expired) {
-			return xerrors.New("expired")
-		}
-
-		time.Sleep(time.Millisecond * 1000)
-	}
-
-	return nil
 }
